add: add login and register logic

+++ database/migrations/2023_06_03_150048_add_user_id_to_events.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('events', function (Blueprint $table) {
-            $table->unsignedInteger('user_id')->after('id');
+            $table->unsignedBigInteger('user_id')->after('id');
 
             $table->foreign('user_id')->references('id')->on('users')->onUpdate('cascade')->onDelete('cascade');
         });
